requirements.txt
# VALORANT Matchup Scouting Assistant - Dependencies
# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Templating
jinja2>=3.1.0

# Environment Variables
python-dotenv>=1.0.0

# Caching
cachetools>=5.3.0

# Date/Time handling
python-dateutil>=2.8.0
